requirements.txt
openai
google-genai
nltk
fastapi
pydantic
dotenv
streamlit